moved expectedReturns calculations into own file
