+++ src/components/staking/PageValidator.vue
@@ -204,7 +204,8 @@ import moment from "moment"
 import { calculateTokens } from "scripts/common"
 import { mapGetters } from "vuex"
 import num, { atoms, viewDenom, shortDecimals } from "scripts/num"
-import { formatBech32, expectedReturns } from "src/filters"
+import { formatBech32 } from "src/filters"
+import { expectedReturns } from "scripts/returns"
 import TmBtn from "common/TmBtn"
 import { ratToBigNumber } from "scripts/common"
 import DelegationModal from "src/ActionModal/components/DelegationModal"
